pyproject.toml
[tool.poetry]
name = "ip-godovich"
version = "0.1.0"
description = "rating system for comics"
authors = ["SergeiNaum <sergei.naumencko@yandex.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"
drf-spectacular = "^0.27.1"
psycopg2-binary = "^2.9.9"
redis = "^5.0.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
black = "^24.1.1"
ipython = "^8.5.0"
flower = "^2.0.1"
django-extensions = "^3.2.3"
werkzeug = "^3.0.1"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
